requirements.txt
auditpol==1.1.0
bcrypt==4.0.1
cffi==1.15.1
colorama==0.4.6
cryptography==40.0.1
et-xmlfile==1.1.0
numpy==1.24.2
openpyxl==3.1.2
pandas==2.0.0
paramiko==3.1.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
tabulate==0.9.0
tzdata==2023.3
